Extract <TestRunSummary />
